Option to exclude unreachable branches
Adding new option, '--exclude-unreachable-branches' that
allows to exclude branches in certain lines from coverage report.
This is intended to solve the problem of reporting uncovered
branches on some compiler-generated code which is essentially dead,
that is to say, cannot be exercised by testing suites.

More details and explanation on sample code:
http://stackoverflow.com/questions/7199360/