If given, options.output is relative to start directory.
If we enter a relative path with --output, we want it to be relative to
the directory we stand in when we run gcovr.

Because we call os.chdir in the code, the easiest way to guarantee this
is to make the path absolute before we start calling os.chdir.