Fix gcov '/'s breaking --html-details on Windows
The gcov reports (from which gcovr output is generated) use '/' as the
path separator in filenames, regardless of the path separator used by
the operating system. This breaks attempts to read files located in
subdirectories while generating --html-details reports on Windows.

This commit fixes the issue by replacing '/'s present in the gcov
filename with Python's OS-sensitive os.sep character.