re-enable HTML reports under Python3
In commit 1cc987a23, the `commonpath()` function was added, which quite
sensibly tries to figure out real path names. Unfortunately, it uses
`iterools.izip()` which is not available under Python3. Instead, that
functionality is now provided by the `zip()` builtin.

We therefore define an `izip()` function as an alias for
`itertools.izip()` if that is available, or else use `zip()`.

Since `commonpath()` is used in HTML reports, it was previously
impossible to create HTML reports with GCOVR under Python3.
To prevent such problems in the future, the existing HTML tests should
be updated and enabled.