fix HTML tests
The HTML tests are naturally rather fragile, since lots of CSS is copied
into every file. However, they are quite valuable as they exercise
important code paths. For example, this would have found issues #163
or #168. By enabling these tests, we can avoid such problems in the
future.

Summary of changes:

**`scripts/gcovr`:**
removed some trailing whitespace in the HTML template.

**`gcovr/tests/test_gcovr.py`:**
normalize HTML files so that they can be compared. In particular, we
remove the exact version string, and the report creation date.

**`gcovr/tests/*/reference/coverage.html`:**
update the HTML files to match the current structure. This is mostly
removing trailing whitespace, and tracking changes to the CSS. If there
was some discrepancy in the content, the TXT and XML references were
assumed to be up to date and fresh values taken from there.

**`gcovr/tests/linked/reference/coverage.html`:**
update as with the other tests. However, this exposed significant
changes to the path name handling. Commit 1cc987a23 changed
`commonprefix()` to `commonpath()` to figure out better path names. This
function is used only by the HTML test, which therefore reports physical
paths. The HTML reference has been updated to use physical paths. E.g.
since the test case is constructed on the fly from the “nested” test,
the source root dir is `../nested/subdir`. Note that the XML and TXT
reports still use symlinks in paths, but that is another problem.