don't --gcov-ignore-parse-errors by default
The new `--gcov-ignore-parse-errors` flag controls whether the
`gcovr.gcov.GcovParser.parse_all_lines()` method throws on errors.

If the flag is enabled, gcovr will

 - print all offending lines,
 - print all caught exceptions,
 - and continue processing.

If the flag is not enabled (the default), gcovr does not continue but

 - displays an error message that the flag can be enabled,
 - then rethrows the first exception if one was caught
 - or `sys.exit(1)` if no exceptions were caught.

In all cases, the errors are processed only *after* a the complete file
has been parsed.

To implement this change, the `check_*()` methods of `GcovParser` were
pulled into the `parse_all_lines()` method.