docs: disable section numbering
The section numbers were used in the original user guide, but now have
little value when using sphinx-docs. Numbering has to be disabled
because this will render the changelog confusingly. E.g.:

    3.1. 3.4 (12 February 2018)

where it should read:

    3.4 (12 February 2018)

There doesn't seem to be a clear way to disable toctree-numbering for a
single document.