Drop support for Python 2.6
I've previously advocated for keeping 2.6-compatibility, but have
changed my mind. There are lots of small inconveniences when restricted
to the 2.6/3.x common subset. In particular, this drastically restricts
the choice of upstream modules.

Old versions of gcovr will of course continue to work.
This also means there will never be a gcovr version that works well on
Windows with Python 2.6, but since that OS doesn't bundle Python such a
version lock-in is less likely.