link to the Gitter chat room
Having a chat is desirable because real-time discussions are sometimes
much more efficient than typing up issue comments. I tested this by
walking a first-time contributor through Git and the pull request
process.

Gitter is a chat service with good GitHub integration. The chat rooms
are fairly open and can be viewed without having to log in. For
logged-in users, there is an IRC bridge.