Initial pass at a separate configuration module
This creates an GcovrConfigOption class to handle the configuration. It
was designed with argparse in mind but it can also be adapted to other
systems. All config options are stored in the GCOVR_CONFIG_OPTIONS
list.

Initially there was also a class representing parsed values, so that we
can check whether a value has been set. Instead, argparse.SUPPRESS is
used to avoid a default assignment of None when the option has not been
provided.