support `--root ..` style invocation
In an out of source build, the compiler will usually be called from
within the build directory:

    cd build; gcc ../src/foo.c -o foo.o

This means gcovr should be called in the same way:

    cd build; gcovr --root ../src .

(Note that the root should point to the source code root. In this case
the root and the search paths differ, so an explicit search path "." has
to be provided. The `--object-directory` option can also be used to
provide a search path, but this also changes some heuristics.)

The oos and oos2 tests were previously identical (oops). The oos2 test
was changed to model the above scenario.

This commit makes that work, without breaking tricky tests such as
"linked":

The `--root` is turned into an `DirectoryPrefixFilter`. This filter will
now use the realpath() of the root, and will normpath() the path to be
matched. This will not break the match if the path below the root
contains symlinks, but does remove any `..` segments from the path that
might have been caused by the `gcc ../` style invocation.

(Possible failure mode if the root path contains symlinks?)

The normpath() version is also used as the file key in the coverage data
structures, thus leading to correct and concise reports that can remove
the root path from the front.

Fixes #294.