Fix overzealous substitution of root folder
If the root folder occurs somewhere in the file name, it is replaced by the empty string by `root_filter.sub('', filename)`, resulting in an invalid file name. Consider for example the transformation
```
/src/libs/libjhtml/src/libjhtml/DataUri.cc -> libs/libjhtmllibjhtml/DataUri.cc
```
when `--root /src` is given.

This PR fixes the problem by only substituting the root folder at the start of the file name.