make CI run successfully on all branches
This PR contains two changes.

First, the **milestone check** only makes sense on pull requests, so it's now
limited to events of that type. Previously, this would cause the CI to fail on
the master branch.

Second, running the CI is also useful on **branches other than master**. In
particular, this lets contributors run CI on their forked repositories.