Add --timestamp option for reproducible reports.
Certain report formats include a timestamp (HTML, Coveralls, Cobertura). It may
however be desirable to attribute the report to a different time:

* e.g. date of a commit instead of time when gcovr was run
* CI server might be running with wrong timezone
* the gcovr test suite

Currently, the gcovr test suite scrubs timestamps from the generated files in
order to enable comparison. In the future, this option could be used instead.
The HTML-example test was already modified to use fixed timestamps in order to
avoid unnecessary Git diffs (the html-details example outputs are committed).

The supported date formats for the --timestamp option are the Unix epoch and
formats that are like RFC-3339 ("ISO dates" without the more insane features).
Time zones are optional.

This should be able to directly understand the outputs from the following GNU
date commands:

    date --rfc-3339=seconds
    date --iso-8601=seconds
    date +%s

It will also understand the following date formats used by Git:

    %aI, %cI, --date=iso-strict, --date=iso8601-strict
    %at, %ct, --date=unix
    %ad, %cd (when appropriate --date format is selected)

Not supported is the RFC 2822 / RFC 5322 format known from emails (e.g.
"Mon, 27 Dec 2021 19:09:44 +0100"). Also unsupported are timestamps with
sub-second accuracy.

While the timestamp parser will guess the format (Epoch or RFC-3339) correctly,
it is also possible to specify a particular format using a prefix:

    gcovr --html --timestamp=epoch:1640606727
    gcovr --html --timestamp=@1640606727
    gcovr --html --timestamp="rfc3339:2021-12-27 13:05:27"

That is, the supported prefixes are:

prefix     description
---------- --------------------------------
`@`        Unix timestamps, matches GNU date
`epoch:`   Unix timestamps
`rfc3339:` RFC-3339 like timestamps

This will enable backwards-compatible future evolution of the supported formats.
The prefixes are not yet mentioned in the `--help` message.