add docs for timestamp feature
The docs were added to a separate page instead of making the Guide file even
larger. Unfortunately, the toctree to link the pages had to be added to a
separate heading due to Sphinx limitations.

The documentation includes a number of tests. These just ensure that various
timestamps are parsed successfully and provide no further checks. If the tests
are executed from within a Git repository, the HEAD timestamp is used for
testing. This could lead to some fragility, but I'm reasonably confident that
the timestamp parsing will always succeed.

Unfortunately these tests are pretty slow. In the future, it might be worth
considering strategies to improve gcovr startup time.