Run release-checklist only in the "deploy" GH Workflow
The release checklist was previously executed as part of the "deploy" and "test"
workflow. In the test configuration, it would demote certain errors to warnings.

However, pull requests run both the test and deploy workflows. Thus, problems
that should only be treated as warnings still caused CI to fail.

The solution was to perform the checks only as part of the "deploy" workflow,
and to configure it to demote errors when not actually doing a release.