docs: format list of available output formats in README more nicely
The README is the de-facto landing page of the project. It includes a nice list
of supported output formats, along with a link to the command line option.

UNFORTUNATELY, this link uses Sphinx syntax that is not supported by the normal
RST (docutils) renderer. This leads to a really ugly view on GitHub with raw RST
syntax. For example:

    * :option:`--html<gcovr --html>`: HTML summaries

SOLUTION: the list is now formatted as a table consisting of RST substitution
references. The table is written in the README file but also included on the
website. Each variant defines these substitutions differently:

* The README version renders the CLI flags as verbatim text without links.
  Links to the guides go to the stable version of the website.

* The website version renders the flags and guide links via the expected Sphinx
  syntax.

Drawback of this change is that it will take more effort to maintain these
links. It's also less than ideal to link to the stable version, but that seems
to be the best approach when considering that outside users are likely to follow
these links.