introduce --cobertura option as alias for --xml
The "XML" format is a confusing name since it could relate to both the Cobertura
and Sonarqube output. This commit adds the --cobertura option as a less
confusing alternative. The documentation was updated throughout to nudge the
user towards this option.

As a side effect, this changes the config file key from `xml` to `cobertura`.

This commit does not deprecate the old -x/--xml/--xml-pretty option names,
but only because it is not reasonably possible to emit a deprecation message.
In contrast, the configuration key had to be renamed since config keys cannot
have aliases.

The following approaches for generating a deprecation warning with argparse were
considered but ultimately discarded:

* wrapping the "type" conversion function for the argument.
  This doesn't work because a "type" function is not supported for some "action"
  types.

* implementing custom "action" types.
  This would require re-implementing the argparse._StoreAction classes and so
  on, which is a lot of work for limited gain. While argparse does offer a
  stable interface for custom actions, this would require further
  integration/reimplementation of argparse details.

* writing a custom argparse replacement.
  This is not insane since value conversion is already highly customized,
  and all that argparse provides is associating --option names with their
  values. However, we cannot get rid of argparse without also rewriting the
  Sphinx-Autoprogram extension.