extract coverage merging functions into new "merging" module
Previously, coverage data was updated in-place with

    coverage.update(new_coverage)

However, this has some problems:

* It cannot deal properly with the DecisionCoverage cases,
  since the type checker (correctly!) refuses to allow updating of different
  types.

* Updating in-place is mutable state, and that is conceptually complicated.

* It is increasingly clear that merging coverage is not just a convenience
  function, but that it is business logic in its own right. In particular, issue
  #586 suggests that in the future, merging will have to depend on command line
  options. This should not be done directly in "gcovr.coverage" since that must
  be a leaf module without further dependencies.

So now a more functional interface is implemented:

    from .merging import merge_file

    coverage = merge_file(coverage, new_coverage)

This might reuse the contents of the input objects, so that the argument objects
must not be used afterwards.

To support this change, some of the coverage processing functions in
"gcovr.writer.json" and "gcovr.__main__" were also updated to return coverage
instead of updating it in-place.

While re-implementing branch coverage merging, it became clear that the
tri-state booleans None/False/True for the "fallthrough" and "throw" fields was
ill-defined. Whereas the gcov-parser would only produce None/True, the JSON
reader/writer coerced these to a False/True bool. Here, None and False mean
"unknown". But the old branch coverage update() function only checked whether
the value was None, meaning that coverage data loaded via "--add-tracefile"
could not be merged successfully. This was fixed by keeping these fields as
ordinary booleans: True if known, False if unknown.