migrate gcovr.writer.cobertura to new coverage data model
This unearthed two bugs, which are kept for now. They should be fixed in a
separate PR for better visibility.

Function coverage in Coberatura reports
---------------------------------------

The Cobertura writer does its own coverage data aggregation. When calculating
package-level function coverage, it would update the package-level metric
by *overwriting* it with the coverage from the last file, instead of *adding*
the counts.

Branch and decision coverage aggregation
----------------------------------------

For calculating line coverage, only lines that are explicitly covered or
uncovered are considered.

For calculating other kinds of coverage metrics (in particular branch coverage),
all lines are considered regardless of whether that line would be considered for
line coverage. Changing this would lead to substantially different coverage
metrics, so it is kept for now.