gcovr.coverage: remove line(), branch(), function() mutating methods
These methods *look* like they are ordinary accessors, but actually create the
coverage object if it doesn't already exist. This causes problems. (A) This
behavior was not expected in some parts of the code, especially the
decision-analysis. (B) The default LineCoverage object does not behave like a
neutral element for the merge_line() operation due to issues with the "noncode"
flag.

This commit DOESN'T FIX those issues, but makes them more visible. Mutation of
the coverage data is now done almost entirely via the "gcovr.merging" module.

To assist with this, the "gcovr.merging" module now contains "insert()"
functions which add a single element into the next-larger structure, for example
"insert_function_coverage()" inserts FunctionCoverage into FileCoverage. In
order to ensure bug-compatibility, Uses of the "line()" method are removed with
the more clearly named "get_or_create_line_coverage()" function.

Merging/updating function coverage is special in that it will only merge data
when the functions are defined on the same line, yet the gcov-parser will update
the line number if the function is redefined. To account for this (and future
developments), all "insert_*()" and "merge_*()" functions now take an optional
"MergeOption" parameter.