more flexible gcov file parsing (fixes #621)
Per the GCC documentation, source lines with zero coverage look like:

```
 #####:  123: some_code();
```

That is, the count column contains a 5-character marker and a leading space,
for a total width of 6 characters.

Some other tools seem to generate files with a SIX (6) character marker,
breaking the parser.

This commit allows any marker that is at least 5 chars long.