handle HTML themes in CSS only; restore tab-size feature
Previously, colors for the themes were specified in Python code.

Now, the colors for all themes are listed in the CSS styles. The current theme
is selected via a "theme-green" or "theme-blue" class on the "<html>" element.

In the future, this would make it possible to switch themes at runtime with a
bit of JavaScript that changes the classes in the DOM.

While making this change, it was discovered that the "tab_size" variable wasn't
used properly. This fixes the regression in
b186a6f04f8b62a8f88c6fb0f60af36af778fe34 / #381.