remove "noncode" concept for LineCoverage
Gcovr previously had a concept of "noncode" lines that do not carry meaningful
coverage but nevertheless are represented with a LineCoverage object. This was
complex and confusing.

Now, noncode status is represented simply by the absence of a matching
LineCoverage object.

This does slightly change the JSON output, but does not affect any other output
formats.