When invoked with --calls, show call coverage
Show them in summary writer. Total number of function calls and number
of calls executed.

HTML view: Show number of calls for each source file and the number of
them executed.

HTML view: In the source view, show the number of calls and executed calls for each
line in their own column.