Use package manager to install ninja because of deprecation warning in seanmiddleditch/gha-setup-ninja@master.
There is a warning and no response to the PR from the developer. To remove this warning we install it now with the package manager.

Node.js 12 actions are deprecated. For more information see: https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/. Please update the following actions to use Node.js 16: seanmiddleditch/gha-setup-ninja@master