Fix nested report if only one directory in root is present. (#750)
- Add comments and to collapse code and remove not needed endless loop.
The loop is always executed only once because because the first statement in the block isn't changed in a second iteration.
- Collapse also the data in the root.
- Use `pop(key)` instead of copying the whole dict ignoring the value.
- Fix evaluating of common path in nested HTML report.
The root_filter wasn't applied to the root directory in HTML report since there is no trailing path separator at the end. Now the directories always have a trailing path separator which is removed in the report generation. To get the common path part the root directory is excluded.
- Replace recursive generation of directory stats with a loop.
- Use a object instead of a dataclass and remove the recursive function and the tests for it.
- Improve detection of root directory.