Add merge for missing properties (#1069)
* Refactor merging of properties
* Refactor error context handling
* Add merge for source and destination block IDs
* Remove condition merge mode
* Use tuples as keys for line and branch coverage

There can be several coverage entries for a line due to the existence of
lambda functions.
To merge them correct the key for the line is extended
by the function name or an empty string if we don't have a function
name. The key for the branch coverage is extended with the source and
the destination block id

We also have function information in legacy text format. For each
function a separate block is generated in the text file. If Lambda
functions are used the line numbers are repeated with different values.
Therefore we can provide functions statistics for older GCC versions,
too.