Fix merging off different reported data across GCOV data files (#1092)
The marking of the coverage data is based on the properties of the item. If, e.g. the branch number,
the block IDs or the length of the conditions are different the data isn't merged but kept separate.

While merging different input files the same branchno can be defined
several times. Sort the branches and enumerate a new branchno for LCOV
and Coveralls report.