Improve pytest execution to be usable in IDE
Move determination of CXX and GCOV from noxfile to conftest and only
rely on environment variable CC in conftest. Log all versions in pytest
instead of noxfile.

Add possibility to copy specific sources only. If there is a file or
directory in the source folder with the test ID as name, this file is
only used.