Fix handling of lines after function specialisation (#1189)
Add a stack to handle function specializations. If a line separator is
found it is ignored if we are not in a specialization, else we switch
back to the function before the specialization. If we find the function
specialization name we store the current state and start with a empty
parser state.